requirements.txt
Flask
scikit-learn
PyMuPDF
python-docx
spacy